[WIP] Many changes are afoot, see comments in code, read more:
This commit is not in any working state, the code has not been executed.
The main goal of this WIP version is the aggregation of the Jira in the
simplest form, using users as an aggregation clause.

Also: the goal is to export a single worksheet first. Chunking will be
added later, if deemed necessary. (since there is always the possibility
of running the program multiple times with different time ranges)